+++ R/sparse_matrix.R
@@ -7,6 +7,7 @@
 #' @importFrom methods is
 #'
 #' @examples
+#' library(Matrix)
 #' is_sparse(matrix(1:10)) # FALSE
 #' is_sparse(Matrix::rsparsematrix(100, 200, .01)) # TRUE
 is_sparse <- function(x) {